requirements/base.txt
requests
beautifulsoup4
slug
python-dotenv
schedule
feedparser
python-telegram-bot
finviz
exif
dataset
tweepy
praw
pytube
vaderSentiment
openai
py-executable-checklist==1.3.1
peewee
pyppeteer
rich
gTTS
